feat: consolidate web utilities
