--- packages/image-link-generator/tsconfig.json
@@ -1,12 +1,0 @@
-{
-	"extends": "../../config/tsconfig.base.json",
-	"compilerOptions": {
-		"rootDir": ".",
-		"outDir": "dist",
-		"composite": true,
-		"declaration": true,
-		"lib": ["ES2022", "DOM"],
-		"skipLibCheck": true
-	},
-	"include": ["src/**/*", "tests/**/*"]
-}
