+++ packages/tests/package.json
@@ -36,7 +36,8 @@
         "@promethean/markdown": "workspace:*",
         "@promethean/parity": "workspace:*",
         "@promethean/stream": "workspace:*",
-        "@promethean/web": "workspace:*",
+        "@promethean/test-utils": "workspace:*",
+        "@promethean/web-utils": "workspace:*",
         "@types/javascript-time-ago": "^2.5.0",
         "@types/unist": "^3.0.3",
         "body-parser": "^1.20.2",
