+++ packages/tests/src/stream-title.test.ts
@@ -3,7 +3,7 @@ import { EventEmitter } from 'node:events';
 import test from 'ava';
 import ollama, { type Message } from 'ollama';
 
-import { sleep } from '@promethean/test-utils/sleep';
+import { sleep } from '@promethean/test-utils/sleep.js';
 import {
     generateTwitchStreamTitle,
     generateAndStoreTitle,
