--- packages/web-core/ava.config.mjs
@@ -1,1 +1,0 @@
-export { default } from '../../../config/ava.config.base.mjs';
