--- packages/web-core/src/shims.d.ts
@@ -1,1 +1,0 @@
-declare module 'fastify';
