+++ packages/web-utils/ava.config.mjs
@@ -1,0 +1,1 @@
+export { default } from '../../config/ava.config.mjs';
