+++ packages/web-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@promethean/image-link-generator",
+  "name": "@promethean/web-utils",
   "version": "0.0.1",
   "private": true,
   "type": "module",
@@ -14,9 +14,7 @@
     "./dist/*": "./dist/*",
     "./*": "./dist/*"
   },
-  "files": [
-    "dist"
-  ],
+  "files": ["dist"],
   "scripts": {
     "build": "tsc",
     "clean": "rimraf dist",
@@ -28,6 +26,9 @@
   },
   "module": "dist/index.js",
   "dependencies": {
+    "cheerio": "^1.0.0-rc.12",
+    "zod": "^3.23.8",
     "@xenova/transformers": "^2.17.2"
-  }
+  },
+  "devDependencies": {}
 }
