--- packages/webcrawler/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "webcrawler",
-  "version": "0.0.1",
-  "type": "module",
-  "main": "dist/index.js",
-  "scripts": {
-    "build": "tsc",
-    "test": "pnpm run build && ava --config ../../config/ava.config.mjs dist/test/**/*.js",
-    "lint": "pnpm exec biome lint .",
-    "format": "pnpm exec biome format --write ."
-  },
-  "dependencies": {
-    "@promethean/web": "file:../web",
-    "cheerio": "^1.0.0-rc.12"
-  }
-}
