--- packages/webcrawler/src/shims.d.ts
@@ -1,1 +1,0 @@
-declare module '@promethean/web';
